+++ cmd/server/main.go
@@ -27,6 +27,7 @@ func main() {
 	mux.Handle("GET /docs/components/card", templ.Handler(pages.Card()))
 	mux.Handle("GET /docs/components/input", templ.Handler(pages.Input()))
 	mux.Handle("GET /docs/components/accordion", templ.Handler(pages.Accordion()))
+	mux.Handle("GET /docs/components/datepicker", templ.Handler(pages.Datepicker()))
 
 	fmt.Println("Server is running on http://localhost:8090")
 	http.ListenAndServe(":8090", mux)