+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added: Input component
 - Added: Accordion component
+- Added: Datepicker component
 - Changed: Add button type prop
 - Changed: Use own input component on tabs example
 - Changed: Update all current components with library comments